requirements.txt
requests==2.32.5
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
